requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
pytest==8.3.3
pytest-flask==1.3.0
